requirements.txt
alembic==1.5.3
apache-airflow==2.0.0
apache-airflow-providers-ftp==1.0.0
apache-airflow-providers-http==1.0.0
apache-airflow-providers-imap==1.0.0
apache-airflow-providers-sqlite==1.0.0
apispec==3.3.2
appdirs==1.4.3
argcomplete==1.12.2
attrs==20.3.0
Babel==2.9.0
CacheControl==0.12.6
cached-property==1.5.2
cattrs==1.2.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.3
colorlog==4.0.2
commonmark==0.9.1
connexion==2.7.0
contextlib2==0.6.0
croniter==0.3.37
cryptography==3.3.1
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
graphviz==0.16
gunicorn==19.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipaddr==2.2.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
msgpack==0.6.2
natsort==7.1.1
numpy==1.20.0
openapi-spec-validator==0.2.9
packaging==20.3
pandas==1.2.1
pendulum==2.1.2
pep517==0.8.2
prison==0.1.3
progress==1.5
psutil==5.8.0
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.22.0
retrying==1.3.3
rich==9.2.0
setproctitle==1.2.2
six==1.14.0
SQLAlchemy==1.3.23
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.13.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
